Create tracking to only detect hands after 2 seconds
has multiple fundamental issues right now. it is able to track the person from a previous frame based off of change in distance
and then if the hand is raised it will count towards basically this streak and if that streak reads 60(2 seconds with 30fps), 
then it will identify it as the hand being raised. However, if just one frame doesn't identify the raised hand(happens pretty often), then it will reset the streak making it a lot harder for a person's raised hand to be detected. Another thing, which isn't that hard but isn't in here yet is just removing the dot when the hand goes down because it keeps the streak, meaning that even after detection it will count in addition to that original 60 frames.